// src/game/TimePilot.types.ts
export type Heading = number;

export interface Coordinates {
  posX: number;
  posY: number;
}

export interface PositionedRadius extends Coordinates {
  radius: number;
}

export interface SpriteFrame extends Coordinates {
  frameWidth: number;
  frameHeight: number;
  frameX: number;
  frameY: number;
}

export interface RenderTextOptions {
  align?: CanvasTextAlign;
  valign?: CanvasTextBaseline;
  size?: number;
  color?: string;
  font?: string;
  stroke?: string | false;
  strokeWidth?: number;
}

export interface CircleOptions {
  backgroundColor?: string;
  borderColor?: string | false;
  borderWidth?: number;
  strokeColor?: string;
}

export interface AssetProgress {
  loaded: number;
  remaining: number;
}

export interface PlayerData extends Coordinates {
  isAlive: boolean;
  deathTick: number | false;
  isFiring: boolean;
  isShooting?: boolean;
  heading: Heading;
  newHeading: Heading | false;
  exploading: number;
  continues: number;
  lives: number;
  score: number;
  level: number;
  removeMe?: boolean;
}

export interface BulletData extends Coordinates {
  heading: Heading;
  size: number;
  velocity: number;
  color: string;
}

export interface EnemyData extends Coordinates {
  heading: Heading;
  level: number;
  deathTick: number | false;
  tickOffset: number;
}

export interface PropData extends Coordinates {
  level: number;
  type: number;
  layer: number;
}

export interface BonusData extends Coordinates {
  level: number;
  layer: number;
  removeMe: boolean;
}

export interface SpriteImage extends HTMLImageElement {
  frameWidth?: number;
  frameHeight?: number;
  frameX?: number;
  frameY?: number;
}

export interface ControllerCommands {
  restart?: () => void;
  pause?: () => void;
}

export interface Controller {
  disconnect?: () => void;
}

export interface GameArenaInstance extends Coordinates {
  width: number;
  height: number;
  updatePosition: (posX: number, posY: number) => void;
  resize: (width?: number, height?: number) => void;
  getContext: (
    dimensions?: "2D" | "2d" | "3D" | "3d" | 2 | 3
  ) => CanvasRenderingContext2D | WebGLRenderingContext;
  enterFullScreen: () => void;
  exitFullScreen: () => void;
  toggleFullScreen: () => void;
  setBackgroundColor: (color: string) => void;
  clear: () => void;
  registerAssets: (assets: string | string[]) => void;
  preloadAssets: (callback?: (progress: AssetProgress) => void) => void;
  renderText: (
    message: string | number,
    startPosX?: number,
    startPosY?: number,
    options?: RenderTextOptions
  ) => void;
  renderSprite: (sprite: CanvasImageSource, spriteData: SpriteFrame) => void;
  drawCircle: (
    posX: number,
    posY: number,
    radius: number,
    options?: CircleOptions
  ) => void;
  drawDebugGrid: (widthSpace?: number, heightSpace?: number) => void;
}
// ...
export interface PlayerInstance {
  getData(): PlayerData;
  getData<K extends keyof PlayerData>(key: K): PlayerData[K] | undefined;
  setData: <K extends keyof PlayerData>(
    key: K,
    value: PlayerData[K],
    isLastKnownGood?: boolean
  ) => boolean;
  resetData: () => void;
  reposition: () => void;
  rotate: () => void;
  startShooting: () => void;
  stopShooting: () => void;
  shoot: () => void;
  render: () => void;
  kill: () => void;
}